Add files via upload
First Download the Code And Extract in a Folder and after that ,
follow these Steps :-
Open Google Chrome and navigate to the Extensions page. You can do this by typing chrome://extensions in the address bar or by going to the menu (three vertical dots) and selecting "More tools" > "Extensions."

On the Extensions page, make sure the "Developer mode" option is enabled. You'll find it located in the top-right corner of the page.

Once you have enabled Developer mode, a new set of buttons will appear on the page. Click on the "Load unpacked" button.

A file browser window will open. Navigate to the folder where you saved your extension files, select the folder, and click "Select" or "Open."

Chrome will load the unpacked extension, and it will appear in the list of installed extensions on the Extensions page.